FSSubSystem creation
BZ: 27903

The FS Subsystem can be used to write into a file of a
filesystem (sysfs as well as regular fs).

An XML configuration file shall be populated in order
to use this subsystem. This configuration file shall
describe the location of the output file and the type of
parameter to be used by the PFW (however, parameters are
always written as strings into the output file):
- Location of the file can be set up using Mapping
properties of the Parameters in the XML file (setting
Directory and File values).
- The type of the parameter is used by the PFW to check
validity of its values (range checking).

Change-Id: I8a15593304a3869d0b95ee1e7a2af401244ff8f6
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>