Clean the _POSIX_C_SOURCE compilation flag
BZ: 105425

__POSIX_VISIBLE will be set to 198808 if _POSIX_SOURCE is defined and
_POSIX_C_SOURCE is not.  This will lead to some bionic/libm
functions invisible to this module.  Not defining any of them would
lead to __POSIX_VISIBLE being set to a safe, default value.

However, either _POSIX_SOURCE or _POSIX_C_SOURCE need to be defined
when using alsa-lib, or it would lead to redefining timeval and
timespec.

So, we have to use _POSIX_C_SOURCE=200809 (the default value if it
wasn't set) for projects using ALSA and we can remove the
_POSIX_SOURCE flags for other projects, that were using it for legacy
reason.

Change-Id: I961b571dd7519ae047c01d6577ee6b6379fa1e5b
Signed-off-by: Du Chenyang <chenyang.du@intel.com>
Signed-off-by: David Wagner <david.wagner@intel.com>