Use low level write API for FS plugin
BZ: 110030

Parameter framework FS plugin currently uses
iostream to write to file system. The underlying STL
library implements the write in a weird way where the
1st character is written then the remaining part of
the string is written. This is not compatible with
sysfs which requires the string to be written at once.

Need to use low level write API to ensure the string
is written at once.

Change-Id: I7060bcf3e40e793326f72e02bbd672fd407d717e
Signed-off-by: Christophe Pavot <christophe.pavot@intel.com>