Change the element builder semantic
BZ: 122982

To parse XML, the PFW uses factories containing a map of XML tags and
their associated builder builder. The builders used to contain the XML
tag of the XML element that it will use in the context of the factory.
A builder and an XML tag are already linked by the factory, thus the tag
member in the builder is redundant.

Remove the XML tag builder attribute and provide it on factory filling.

Change-Id: I827c3b34dffb23bc850316d0b092b31c21987ac2
Signed-off-by: Kevin Rocard <kevinx.rocard@intel.com>