Add a CMake Makefile
You need to have the parameter-framework core libraries and headers installed.

Run "cmake ." for generating the Makefile, then "make" to compile;
"make install" to install.

If you have installed the parameter-framework in a custom directory, add the
"-DCMAKE_PREFIX_PATH=/path/to/custom/prefix" option to the cmake command.

If you want to install the plugin to a custom location, add the
"-DCMAKE_INSTALL_PREFIX=/path/to/custom/install" option.

Change-Id: I7e75096b25fd870030595a83be6c4d62bf8c11e0
Signed-off-by: David Wagner <david.wagner@intel.com>
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>