Use standard library path for pfw plugin location
The filesystem plugin used to be installed in
/system/lib/parameter-framework-plugins/Fs.
Since the Parameter-framework is now able to load plugins from
standard library locations such as /system/lib(64)/,
there is no need for this specific location anymore.

This patch moves the install directory of libfs-subsystem.so to
the standard library path.

Issue: GMINL-1748

Change-Id: Ib714bc26af4698dcd41aaf56272bf5c1cf2bdf14
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>