Android: Link to shared library xmlSerializer
The PFw as recently removed (this trick was not working on osx)
it's *_include libraries. They were used to include headers
without linking against the libraries in the android build system.

There is two way to include headers from a library:
 - copy the headers (but that would be in contradiction with how the main pfw headers are included)
 - link against the library (chosen solution)
To avoid linking against a static libxmlserializer - ie add it and it's dependency to all plugins - the xmlserializer library is now dynamic.

Link against the dynamic library xmlserializer and no longer use *_include.

Change-Id: I1811a7ab1c5f892adadb8709b8fefa58ebad4542
Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>